requirements.txt
python-telegram-bot[job-queue]==22.1.0
aiogram==3.20.0.post0
flask
python-dotenv
requests
psutil
aiohttp
beautifulsoup4
aiofiles
Pillow
psycopg2-binary
pycryptodome
discord.py
# Grok Image Bot dependencies
playwright
pyTelegramBotAPI
tempmail-python
playwright-stealth
